api/pyproject.toml
[tool.poetry]
name = "image_api"
description = ""
version = "0.1.0"
authors = ["Quentin Gliech <gliech@unistra.fr>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.1"
boto3 = "^1.21.38"
Flask-Cors = "^3.0.10"
Pillow = "^9.1.0"
celery = {extras = ["amqp", "redis"], version = "^5.2.6"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "^0.942"
boto3-stubs = {extras = ["essential"], version = "^1.21.38"}
celery-types = "^0.12.0"

[tool.poetry.scripts]
app = "image_api.web:dev"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
